+++ src/entities/post/api/hooks/usePostQueries.ts
@@ -1,13 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query"
-import { Post, postApi } from "../postApi"
-
-export const postQueryKeys = {
-  all: "posts",
-  lists: () => [postQueryKeys.all, "list"] as const,
-  list: (filters: { limit: number; skip: number }) => [postQueryKeys.lists(), filters] as const,
-  search: (query: string) => [postQueryKeys.all, "search", query] as const,
-  tags: (tag: string) => [postQueryKeys.all, "tag", tag] as const,
-}
+import { postApi, PostsTypes } from "../postApi"
+import { postQueryKeys } from "../postQueryKeys"
 
 export const useGetPosts = (limit: number, skip: number) => {
   return useQuery({
@@ -31,47 +24,7 @@ export const useGetPostsByTag = (tag: string) => {
   })
 }
 
-interface MutationOptions {
-  onSuccess?: () => void
-  onError?: (error: Error) => void
-}
-
-export const useCreatePost = (options: MutationOptions = {}) => {
-  const { onSuccess, onError } = options
-  const queryClient = useQueryClient()
-  const queryKey = postQueryKeys.lists()
-
-  return useMutation({
-    mutationFn: postApi.createPost,
-    onMutate: async (newPost) => {
-      const previousPosts = queryClient.getQueryData(queryKey)
-
-      queryClient.setQueryData(queryKey, (old: Post[] = []) => {
-        const optimisticPost = {
-          ...newPost,
-          id: Date.now(),
-          tags: [],
-          reactions: { likes: 0, dislikes: 0 },
-          views: 0,
-        }
-        return [...old, optimisticPost]
-      })
-
-      return { previousPosts }
-    },
-    onSuccess,
-    onError: (err, _, context) => {
-      queryClient.setQueryData(queryKey, context?.previousPosts)
-      onError?.(err)
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: queryKey })
-    },
-  })
-}
-
-export const useUpdatePost = (options: MutationOptions = {}) => {
-  const { onSuccess, onError } = options
+export const useUpdatePost = () => {
   const queryClient = useQueryClient()
   const queryKey = postQueryKeys.lists()
 
@@ -80,16 +33,14 @@ export const useUpdatePost = (options: MutationOptions = {}) => {
     onMutate: async (updatedPost) => {
       const previousPosts = queryClient.getQueryData(queryKey)
 
-      queryClient.setQueryData(queryKey, (old: Post[] = []) => {
+      queryClient.setQueryData(queryKey, (old: PostsTypes["posts"] = []) => {
         return old.map((post) => (post.id === updatedPost.id ? updatedPost : post))
       })
 
       return { previousPosts }
     },
-    onSuccess,
     onError: (err, _, context) => {
       queryClient.setQueryData(queryKey, context?.previousPosts)
-      onError?.(err)
     },
     onSettled: () => {
       queryClient.invalidateQueries({ queryKey: queryKey })
@@ -97,8 +48,7 @@ export const useUpdatePost = (options: MutationOptions = {}) => {
   })
 }
 
-export const useDeletePost = (options: MutationOptions = {}) => {
-  const { onSuccess, onError } = options
+export const useDeletePost = () => {
   const queryClient = useQueryClient()
   const queryKey = postQueryKeys.lists()
 
@@ -107,16 +57,14 @@ export const useDeletePost = (options: MutationOptions = {}) => {
     onMutate: async (postId) => {
       const previousPosts = queryClient.getQueryData(queryKey)
 
-      queryClient.setQueryData(queryKey, (old: Post[] = []) => {
+      queryClient.setQueryData(queryKey, (old: PostsTypes["posts"] = []) => {
         return old.filter((post) => post.id !== postId)
       })
 
       return { previousPosts }
     },
-    onSuccess,
     onError: (err, _, context) => {
       queryClient.setQueryData(queryKey, context?.previousPosts)
-      onError?.(err)
     },
     onSettled: () => {
       queryClient.invalidateQueries({ queryKey: queryKey })
